// webapp/src/app/core/models/enums.ts
export type ArticleType = "pmc" | "pubmed" | "freetext" | "pdf";

export type ResponseType = "CONSOLE" | "RESPONSE" | "NOT_FOUND" | "ERROR" | "NONE";

/*
export enum ResponseType {
    CONSOLE = "CONSOLE",
    RESPONSE = "RESPONSE",
    NOT_FOUND = "NOT_FOUND",
    ERROR = "ERROR",
    NONE = "NONE"
};
*/